Maximum Product subarray
The idea is to traverse array from left to right keeping two variables minVal and maxVal which represents the minimum and maximum product value till the ith index of the array. Now, if the ith element of the array is negative that means now the values of minVal and maxVal will be swapped as value of maxVal will become minimum by multiplying it with a negative number. Now, compare the minVal and maxVal.
The value of minVal and maxVal depends on the current index element or the product of the current index element and the previous minVal and maxVal respectively.