Maximum path sum in matrix
Approach: Dynamic programming is used to solve the above problem in a recursive way.  

Allot a position at the beginning of the matrix at (0, 0).
Check each next allowed position from the current position and select the path with maximum sum.
Take care of the boundaries of the matrix, i.e, if the position reaches the last row or last column then the only possible choice will be right or downwards respectively.
Use a map to store track of all the visiting positions and before visiting any (i, j), check whether or not the position is visited before.
Update the maximum of all possible paths returned by each recursive calls made.
Go till the position reaches the destination cell, i.e, (n-1.m-1).